pyproject.toml
[tool.poetry]
name = "botlerbot"
version = "0.1.0"
description = ""
authors = ["Yehor Lenko <yehorlenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
ujson = "^5.7.0"
aiohttp = {extras = ["speedups"], version = "^3.8.4"}
python-dotenv = "^1.0.0"
requests = "^2.28.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.1.1"
isort = "^5.12.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
